+++ tests/0100_builder.py
@@ -16,38 +16,32 @@ def test_build_process(caplog, load_json, settings, tmp_path):
     """
     caplog.set_level(logging.DEBUG, logger=__pkgname__)
 
-    # Path to appstack directory
-    basic_appstack = settings.configs_path / "appstack_single_component"
-    dual_appstack = settings.configs_path / "appstack_dual_components"
-
-    # Load application stack configurations
-    basic_appstack_conf = load_json(
-        settings.configs_path / "appstack_single_component/appstack.json"
-    )
-    dual_appstack_conf = load_json(
-        settings.configs_path / "appstack_dual_components/appstack.json"
-    )
-
-    # Load models declarations
-    blog_models = load_json(settings.configs_path / "models_basic_blog.json")
-    cms_models = load_json(settings.configs_path / "models_basic_cms.json")
-
     project = ProjectRegistry()
 
-    # Register applications
-    project.add_application(
-        basic_appstack_conf,
-        basic_appstack,
-        name="Blog app",
-    )
-    project.add_app_models("single-component", blog_models)
-
-    project.add_application(
-        dual_appstack_conf,
-        dual_appstack,
-        name="CMS app",
-    )
-    project.add_app_models("dual-components", cms_models)
+    project.load_configuration({
+        "apps": {
+            "blog": {
+                "name": "Blog app",
+                "destination": "the-blog",
+                "template_dir": settings.configs_path / "appstack_single_component",
+                "declarations": settings.configs_path / "models_basic_blog.json",
+                "appstack": (
+                    settings.configs_path / "appstack_single_component"
+                    / "appstack.json"
+                )
+            },
+            "cms": {
+                "name": "CMS app",
+                "destination": "the-cms",
+                "template_dir": settings.configs_path / "appstack_dual_components",
+                "declarations": settings.configs_path / "models_basic_cms.json",
+                "appstack": (
+                    settings.configs_path / "appstack_dual_components"
+                    / "appstack.json"
+                )
+            },
+        },
+    })
 
     # Run builder processor
     builder = ProjectBuilder(project, tmp_path)
@@ -55,8 +49,8 @@ def test_build_process(caplog, load_json, settings, tmp_path):
 
     # Check expected component directories
     assert sorted(tmp_path.iterdir()) == [
-        tmp_path / "dual-components",
-        tmp_path / "single-component",
+        tmp_path / "the-blog",
+        tmp_path / "the-cms",
     ]
 
     # Check expected module files for all components
@@ -68,9 +62,9 @@ def test_build_process(caplog, load_json, settings, tmp_path):
         ])
 
     assert sorted(built_files) == [
-        "dual-components/plugins/page.py",
-        "dual-components/views/page.py",
-        "single-component/views/__init__.py",
-        "single-component/views/article.py",
-        "single-component/views/blog.py"
+        "the-blog/views/__init__.py",
+        "the-blog/views/article.py",
+        "the-blog/views/blog.py",
+        "the-cms/plugins/page.py",
+        "the-cms/views/page.py",
     ]